First Large Commit
I've written a LOT of code and documentation. Now what I am about to do is a small overhaul in my model api class, so that it stores the frequency and "eigensolution" when solving the main function `generate_eigensolution`. The subsequent functions will have frequency as KEYWORD argument, and will assume that the last used frequency is desired when called without one. This will mean the stored eigensolution is used.